ready-tensor-course/week5/pyproject.toml
[project]
name = "week5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.5.0",
    "langchain-cerebras>=0.5.0",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "pydantic>=2.11.9",
    "pyjokes>=0.8.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
